requirements.txt
openai
python-dotenv
requests
pytest
flake8
black
